backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]>=0.34.0
python-dotenv==1.*
pydantic==2.*
supabase==2.*
websockets>=13.0
ffmpeg-python==0.2.*
httpx==0.*
fastmcp==3.*
tavily-python==0.*
groq==0.*
